work on the saver
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@18636 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/cssaver/cssaver.cpp
@@ -469,7 +469,7 @@ bool csSaver::SaveSectorLights(iSector *s, iDocumentNode *parent)
     centerNode->SetAttributeAsFloat("z", center.z);
 
     //Add the light center node
-    int radius = light->GetInfluenceRadius();
+    float radius = light->GetInfluenceRadius();
     csRef<iDocumentNode> radiusNode = CreateNode(lightNode, "radius");
     radiusNode->CreateNodeBefore(CS_NODE_TEXT)->SetValueAsFloat(radius);
